feat: 오늘의 챌린지, 유저 챌린지 api 추가

+++ challenge/schemas.py
@@ -1,12 +1,11 @@
 from datetime import datetime
-from decimal import Decimal
 
 from ninja import Schema
 
 
 class CoordinateSchema(Schema):
-    latitude: Decimal
-    longitude: Decimal
+    latitude: float
+    longitude: float
 
 
 class AcceptChallengeSchema(Schema):
